agents/test-cases-agent/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "test-cases-agent"
version = "1.0.0"
description = "Test Cases Agent - Generates test case specifications using LLMs"
authors = [
    {name = "QA Platform Team", email = "team@qaplatform.io"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "grpcio-reflection>=1.76.0",
    "grpcio-health-checking>=1.76.0",
    "anthropic>=0.75.0",
    "weaviate-client>=4.18.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.1.0",
    "structlog>=25.5.0",
    "aiohttp>=3.11.0",
    "prometheus-client>=0.21.0",
    "tenacity>=9.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.12.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-protobuf>=5.29.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
test-cases-agent = "test_cases_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["test_cases_agent*"]
namespaces = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/test_cases_agent"]
omit = [
    "*/proto/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    proto
    | migrations
    | .venv
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # `zip()` without an explicit `strict=` parameter
]
exclude = [
    ".venv",
    "proto",
    "__pycache__",
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
strict_equality = true
exclude = [
    "proto/",
    "tests/",
]